src/requirements.txt
ffmpeg_python==0.2.0
langchain==0.3.2
langchain_core==0.3.9
langchain_openai==0.2.2
langchain_pinecone==0.2.0
openai==1.51.1
pinecone==5.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.30.0
pinecone[grpc]
